Restructure app entry points into cmd directory
